requirements.txt
django-tagging
-e git://github.com/pinax/pinax.git#egg=pinax
